Update OrbitalScience.cfg
This is an attempt to bring the DMagic Antennae more in-line with a balanced playthrough experience.
Large dish antenna was changed from 7000Gm to 70Gm
Small was changed to 50Gm

The original values for these antenna were really overpowered for the weight. There was no reason to use anything else if you needed long-range antennae. This is designed to work out as far as Stock Eeloo so it provides a good pro-con for super long range communication by costing a lot of EC and having very low data Tx rates.